+++ source/MaterialXCore/Traversal.cpp
@@ -15,7 +15,6 @@ const Edge NULL_EDGE(nullptr, nullptr, nullptr);
 const TreeIterator NULL_TREE_ITERATOR(nullptr);
 const GraphIterator NULL_GRAPH_ITERATOR(nullptr, nullptr);
 const InheritanceIterator NULL_INHERITANCE_ITERATOR(nullptr);
-const AncestorIterator NULL_ANCESTOR_ITERATOR(nullptr);
 
 //
 // Edge methods
@@ -199,39 +198,16 @@ InheritanceIterator& InheritanceIterator::operator++()
     if (_elem)
     {
         ElementPtr super = _elem->getInheritsFrom();
-
-        // Check for cycles.
-        if (_pathElems.count(super))
+        if (super && super->getCategory() == _elem->getCategory())
         {
-            throw ExceptionFoundCycle("Encountered cycle at element: " + super->asString());
+            // Check for cycles.
+            if (_pathElems.count(super))
+            {
+                throw ExceptionFoundCycle("Encountered cycle at element: " + super->asString());
+            }
+            _pathElems.insert(super);
         }
-
         _elem = super;
-        _pathElems.insert(super);
-    }
-    return *this;
-}
-
-//
-// AncestorIterator methods
-//
-
-const AncestorIterator& AncestorIterator::end()
-{
-    return NULL_ANCESTOR_ITERATOR;
-}
-
-AncestorIterator& AncestorIterator::operator++()
-{
-    if (_holdCount)
-    {
-        _holdCount--;
-        return *this;
-    }
-
-    if (_elem)
-    {
-        _elem = _elem->getParent();
     }
     return *this;
 }