+++ source/MaterialXTest/File.cpp
@@ -34,8 +34,8 @@ TEST_CASE("File system operations", "[file]")
 {
     std::vector<std::string> filenames =
     {
-        "documents/Libraries/mx_stdlib_defs.mtlx",
-        "documents/Examples/MaterialGraphs.mtlx",
+        "documents/Libraries/stdlib_defs.mtlx",
+        "documents/Examples/MaterialBasic.mtlx",
         "documents/Examples/PaintMaterials.mtlx",
     };